+++ fabric/src/main/java/infinituum/labellingcontainers/fabric/LabellingContainersFabric.java
@@ -1,8 +1,8 @@
 package infinituum.labellingcontainers.fabric;
 
 import infinituum.labellingcontainers.LabellingContainers;
-import infinituum.labellingcontainers.fabric.events.BlockPlaceCallback;
 import infinituum.labellingcontainers.fabric.events.BlockPlaceEventHandler;
+import infinituum.labellingcontainers.fabric.events.custom.BlockPlaceCallback;
 import net.fabricmc.api.ModInitializer;
 
 public class LabellingContainersFabric implements ModInitializer {
